Refactors to BLoC pattern
