+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter/material.dart';
 
-import 'package:recipe_app/screens/home_screen.dart';
+import 'src/features/home/home_screen.dart';
 
 void main() async {
   await dotenv.load(fileName: ".env");
