+++ lib/src/features/recipe/presentation/bloc/recipe_event.dart
@@ -1,0 +1,7 @@
+abstract class RecipeEvent {}
+
+class LoadRecipeEvent extends RecipeEvent {
+  final int mealId;
+
+  LoadRecipeEvent(this.mealId);
+}
