+++ pubspec.yaml
@@ -29,7 +29,9 @@ environment:
 # versions available, run `flutter pub outdated`.
 dependencies:
   http: ^0.13.4
+  rxdart: ^0.27.7
   flutter_html: ^2.2.1
+  flutter_bloc: ^8.0.0
   flutter_dotenv: ^5.0.0
   flutter_typeahead: ^4.3.7
   flutter:
